Refactorizar modulo de administracion en sub-modulos enfocados
Se divide `src/modules/admin.js` en `admin-dashboard.js`, `admin-municipalities.js`, `admin-users.js` y `admin-requests.js` para mejorar la mantenibilidad y legibilidad. El archivo principal `admin.js` ahora sirve como orquestador para la inicialización y delegación de eventos.

Cambios:
- Creado `src/modules/admin-dashboard.js`: Carga de estadísticas del dashboard.
- Creado `src/modules/admin-municipalities.js`: Gestión de municipalidades y departamentos.
- Creado `src/modules/admin-users.js`: Gestión de usuarios.
- Creado `src/modules/admin-requests.js`: Gestión de solicitudes de rol.
- Refactorizado `src/modules/admin.js`: Orquestador que importa lo anterior.
